CH-544/CH-629 added ReviewSecuredController and added @Secured
jira [Change review service] https://cozyhometeam.atlassian.net/browse/CH-629?atlOrigin=eyJpIjoiOWU1ODkwNGMxOTRjNDIwNDhjOTBlOGMxMzA1M2E5YzYiLCJwIjoiaiJ9
